fix: broke the dependency cycle
