+++ src/entities/Quiz/index.ts
@@ -1,0 +1,1 @@
+export type { QuizData } from './model/types';
