+++ src/pages/Quizzes/model/cards/index.ts
@@ -2,4 +2,4 @@ export {
   BASE_LIMIT_CARDS_ON_PAGE,
   COUNT_DISPLAY_SKELETON_CARDS,
 } from './constants';
-export { type QuizData, useQuizCards } from './useQuizCards';
+export { useQuizCards } from './useQuizCards';
