+++ src/pages/Quizzes/model/cards/useQuizCards/index.ts
@@ -1,2 +1,1 @@
 export { useQuizCards } from './useQuizCards';
-export { type QuizData } from './useQuizCards.types';
